[clang] template / auto deduction deduces common sugar
This patch implements deducing the common sugar between two types,
and uses it to establish the result when checking deduced template arguments.

This carries over to auto return type deduction as well, which is reworked
to update the deduced type.

The common sugar is calculated by stripping off top-level sugar which differs
between types, crossing over and rebuilding non-sugar type nodes if needed.

Signed-off-by: Matheus Izvekov <mizvekov@gmail.com>

Differential Revision: https://reviews.llvm.org/D111283