[clang] fix uniquing of some TagTypes created from the injected class name
This fixes a bug in the fast path for the creation of TagTypes from
injected class names.

The creation of TagTypes has a fast path which, when there is no
elaboration, uses storage in the declaration itself for memoizing
the resuling type node, instead of using the folding set.

This memoizing would fail when the type was created from the injected
class name, as we would look for the node in the injected declaration
but store it in the non-injected one, so a different type would
be created each time.

This regression was reported here: https://github.com/llvm/llvm-project/pull/147835#discussion_r2299248995
Since this regression was never released, there are no release notes.